requirements.txt
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.8.0
docopt==0.6.2
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hdfs==2.7.3
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.25.2
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.45.3
streamlit==1.42.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
trio==0.28.0
trio-websocket==0.11.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
